requirements.txt
langchain
langchain-community
langchain-openai
langgraph
openai
faiss-cpu
pydantic
python-dotenv
beautifulsoup4
requests
streamlit
wikipedia
langchain-core
langsmith
langchain[openai]
pypdf
pymupdf
streamlit[pdf]
pdfplumber
anthropic